2014.3.15中午
速度闭环成功，继续调试参数

+++ sources/speed_control.h
@@ -3,5 +3,5 @@ void set_motor_highduty(float highdutyA,float highdutyB);
 void Coder_init();
 void get_speed();
 void speed_out(float target_speed);
-extern int Right_motor_speed,Left_motor_speed;//左右轮测的的速度
+extern float Right_motor_speed,Left_motor_speed;//左右轮测的的速度
 extern float Set_speed;
